feat: :sparkles: Render markdown within Annotation components
The annotations can use markdown now. It will be properly displayed in the open annotation and in the article menu.

Nesting annotation components within other components in MDX requires a workaround for the content to render properly. The annotation content needs to be defined in another file and rendered as a JSX component within the Annotation.