feat: :sparkles: Implement customization of section names
The section names in table of contents can now be customized. This can be done via the section_names.json file in meta folder. The key of a section is the index number and the value can either be a string specifying the name of the section or a boolean (false) in case the section name is to be omitted.